[COST-4445] delete KokuDBAccess
